+++ src/App.jsx
@@ -1,21 +1,19 @@
 import { useMediaQuery } from 'react-responsive';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-
 import Header from './component/layout/Header';
 import Footer from './component/layout/Footer';
 import MainPage from './pages/Main';
 import CentralClubPage from './pages/CentralClub';
 import SmallClubPage from './pages/SmallClub';
-import DetailPage from './menu/detail';
+import DetailPage from './pages/DetailPage';
 import LoginPage from './pages/LoginPage';
-import KakaoRedirection from './menu/login/component/kakaoRedirection';
-import BranchSmall from './menu/small_club/branch/branch_small';
+import KakaoRedirection from './component/login/kakaoRedirection';
 import HashTagPage from './pages/HashTag';
-import ReviewWrite from './menu/detail/review_write';
-import Search from './menu/search/search';
+import SearchPage from './pages/SearchPage';
 import SummaryPage from './pages/Summary';
 import BookMarkPage from './menu/bookmark';
 import BranchCentralPage from './pages/BranchCentral';
+import BranchSmallPage from './pages/BranchSmall';
 
 function App() {
     const isPc = useMediaQuery({
@@ -35,13 +33,12 @@ function App() {
                         {/* <Route path="/menu/central_club/detail_page/review_write" element={<ReviewWrite />} /> */}
                         <Route path="/small" element={<SmallClubPage />} />
                         <Route path="/clubs/:clubId" element={<DetailPage />} />
-                        <Route path="/menu/detail/review_write" element={<ReviewWrite />} />
                         <Route path="/central/divisions" element={<BranchCentralPage />} />
                         <Route path="/login" element={<LoginPage />} />
                         <Route path="/v1/auths/oauth/kakao" element={<KakaoRedirection />} />
-                        <Route path="/menu/small_club/branch/branch_small" element={<BranchSmall />} />
+                        <Route path="/small/colleges" element={<BranchSmallPage />} />
                         <Route path="/hastag" element={<HashTagPage />} />
-                        <Route path="/menu/search" element={<Search />} />
+                        <Route path="/search" element={<SearchPage />} />
                         <Route path="/summary" element={<SummaryPage />} />
                         <Route path="/menu/bookmark" element={<BookMarkPage />} />
                     </Routes>
@@ -57,13 +54,12 @@ function App() {
                         <Route path="/central" element={<CentralClubPage />} />
                         <Route path="/small" element={<SmallClubPage />} />
                         <Route path="/clubs/:clubId" element={<DetailPage />} />
-                        <Route path="/menu/detail/review_write" element={<ReviewWrite />} />
                         <Route path="/central/divisions" element={<BranchCentralPage />} />
                         <Route path="/login" element={<LoginPage />} />
                         <Route path="/v1/auths/oauth/kakao" element={<KakaoRedirection />} />
-                        <Route path="/menu/small_club/branch/branch_small" element={<BranchSmall />} />
+                        <Route path="/small/colleges" element={<BranchSmallPage />} />
                         <Route path="/hashtag" element={<HashTagPage />} />
-                        <Route path="/menu/search" element={<Search />} />
+                        <Route path="/search" element={<SearchPage />} />
                         <Route path="/summary" element={<SummaryPage />} />
                         <Route path="/menu/bookmark" element={<BookMarkPage />} />
                     </Routes>
