+++ src/component/hashtag/TagScroll.jsx
@@ -3,7 +3,7 @@ import HashTag from './HashTag';
 
 export default function TagScroll() {
     return (
-        <div className="wrapper">
+        <div className="tagScroll_wrapper">
             <div className="tags_wrapper">
                 <HashTag />
             </div>
