+++ src/component/smallClub/SmallClubProps.jsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from '../centralClub/centralClub.module.css';
+import styles from './smallClubProps.module.css';
 import { LinkItem } from '../branch/BranchCentral';
 
-function HashTagClub({ clubId, imageUrl, clubName, introduction }) {
+//clubid페이지로 이동할 수 있게 수정 완료
+function SmallClubProps({ clubId, clubName, introduction, imageUrl }) {
     return (
         <div className={styles.rectangle}>
-            <LinkItem to={`/menu/detail/${clubId}`}>
-                <img className={styles.image} alt={clubName} src={imageUrl} />
+            <LinkItem to={`/clubs/${clubId}`}>
+                <img src={imageUrl} alt={clubName} className={styles.image} />
+
                 <h3 className={styles.title}>{clubName}</h3>
                 <p className={styles.content}>{introduction}</p>
             </LinkItem>
@@ -15,11 +17,11 @@ function HashTagClub({ clubId, imageUrl, clubName, introduction }) {
     );
 }
 
-HashTagClub.propTypes = {
+SmallClubProps.propTypes = {
     clubId: PropTypes.number.isRequired,
     imageUrl: PropTypes.string.isRequired,
     clubName: PropTypes.string.isRequired,
     introduction: PropTypes.string.isRequired,
 };
 
-export default HashTagClub;
+export default SmallClubProps;
