--- src/menu/detail/_component/reviewBox.css
@@ -1,97 +1,0 @@
-@media screen and (min-width: 768px) {
-    .review_box_container {
-        width: 125%;
-        background-color: white;
-        box-shadow: 0px 0px 4px 0px #0000004d;
-        border-radius: 15px;
-        margin-bottom: 15px;
-        padding: 16px 18px;
-    }
-
-    .review_box_header {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-    }
-
-    .review_box_header p {
-        font-size: 16px;
-        font-weight: 700;
-        margin-right: 7px;
-    }
-
-    .review_box_header span {
-        font-size: 12px;
-        font-weight: 400;
-        color: gray;
-    }
-
-    .review_box_contents {
-        display: grid;
-        /* 한 줄에 세개의 열 */
-        grid-template-columns: repeat(3, 1fr);
-        gap: 5px;
-        margin-top: 7px;
-        /* width: 100%; */
-    }
-
-    .keyword_container {
-        background-color: #9c9c9c26;
-        border-radius: 5px;
-        border: none;
-        padding: 7px 10px;
-        font-size: 13px;
-        font-weight: 400;
-        /* width: 30%;
-    flex: 1; */
-    }
-}
-
-@media screen and (max-width: 768px) {
-    .review_box_container {
-        width: 125%;
-        background-color: white;
-        box-shadow: 0px 0px 4px 0px #0000004d;
-        border-radius: 15px;
-        margin-bottom: 15px;
-        padding: 16px 18px;
-    }
-
-    .review_box_header {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-    }
-
-    .review_box_header p {
-        font-size: 13px;
-        font-weight: 700;
-        margin-right: 7px;
-    }
-
-    .review_box_header span {
-        font-size: 11px;
-        font-weight: 400;
-        color: gray;
-    }
-
-    .review_box_contents {
-        display: grid;
-        /* 한 줄에 세개의 열 */
-        /* grid-template-columns: repeat(3, 1fr);*/
-        gap: 5px;
-        margin-top: 7px;
-        /* width: 100%; */
-    }
-
-    .keyword_container {
-        background-color: #9c9c9c26;
-        border-radius: 5px;
-        border: none;
-        padding: 7px 10px;
-        font-size: 13px;
-        font-weight: 400;
-        /* width: 30%;
-    flex: 1; */
-    }
-}
