--- src/menu/detail/detailPage.css
@@ -1,258 +1,0 @@
-@media screen and (min-width: 768px) {
-    .detail_container {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-
-    .detail_header_container {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        width: 70%;
-        /* height: 40%; */
-    }
-
-    .detail_header_name {
-        display: flex;
-        flex-direction: row;
-        margin-bottom: 10px;
-    }
-
-    .detail_header_name h3 {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        font-size: 20px;
-        font-weight: 500;
-        /* line-height: 20px; */
-        margin-top: 1px;
-    }
-
-    .detail_logo {
-        width: 20%;
-        height: 60%;
-        margin: 12px 24px 20px 22px;
-    }
-
-    .insta_icon {
-        margin-top: 1%;
-        margin-left: 3%;
-        width: 20px;
-        height: 20px;
-    }
-
-    .star_icon {
-        width: 22px;
-        height: 21px;
-        margin-left: 7px;
-    }
-
-    .association_btn {
-        width: 160px;
-        height: 30px;
-        background-color: #7BC8E0;
-        color: white;
-        font-size: 12.5px;
-        font-weight: 400;
-        border-radius: 5px;
-        /* text-align: center; */
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-evenly;
-        padding: 0px 6px 0px 6px;
-    }
-
-    .detail_image {
-        width: 30%;
-        aspect-ratio: 1/1;
-        object-fit: cover;
-        margin-bottom: 2.5%;
-    }
-
-    .detail_tab {
-        position: relative;
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-evenly;
-        width: 70%;
-        height: 36px;
-        background-color: #7BC8E04D;
-    }
-
-    .detail_body {
-        display: flex;
-        flex-direction: column;
-        align-items: left;
-        width: 70%;
-        margin: 2.5% 0 15% 0;
-    }
-
-    .detail_body h3 {
-        width: fit-content;
-        margin: 2.5% 0 2.5% 0;
-        display: inline;
-        box-shadow: inset 0 -15px 0 #7BC8E04D;
-    }
-
-    .detail_body p {
-        margin-top: 1%;
-        margin-left: 3%;
-        width: 50%;
-        text-align: left;
-    }
-
-    .detail_tab button {
-        border: none;
-        background-color: transparent;
-        box-shadow: none;
-        cursor: pointer;
-        width: 10%;
-        height: 105%;
-        font-weight: 500;
-        text-align: center;
-    }
-
-    .detail_tab button.active {
-        border-top: 2.5px solid #7BC8E0;
-        border-bottom: 2.5px solid #7BC8E0;
-        border-radius: 2px;
-    }
-
-    .review_body {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-}
-
-@media screen and (max-width: 768px) {
-    .detail_container {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-
-    .detail_header_container {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        width: 70%;
-        /* height: 40%; */
-    }
-
-    .detail_header_name {
-        display: flex;
-        flex-direction: row;
-        margin-bottom: 10px;
-    }
-
-    .detail_header_name h3 {
-        font-size: 20px;
-        font-weight: 500;
-        line-height: 20px;
-    }
-
-    .detail_logo {
-        width: 30%;
-        height: 60%;
-        margin: 12px 24px 20px 22px;
-    }
-
-    .insta_icon {
-        margin-top: 1%;
-        margin-left: 5%;
-        width: 20px;
-        height: 20px;
-    }
-
-    .star_icon {
-        width: 23px;
-        height: 23px;
-        margin-left: 2px;
-        padding: 2px;
-    }
-
-    .association_btn {
-        width: 150px;
-        height: 30px;
-        background-color: #7BC8E0;
-        color: white;
-        font-size: 11px;
-        font-weight: 400;
-        border-radius: 5px;
-        /* text-align: center; */
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-evenly;
-        padding: 0px 6px 0px 6px;
-    }
-
-    .detail_image {
-        width: 50%;
-        aspect-ratio: 1/1;
-        object-fit: cover;
-        margin-bottom: 2.5%;
-        align-items: left;
-    }
-
-    .detail_tab {
-        position: relative;
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-evenly;
-        width: 70%;
-        height: 36px;
-        background-color: #7BC8E04D;
-    }
-
-    .detail_body {
-        display: flex;
-        flex-direction: column;
-        align-items: left;
-        width: 70%;
-        margin: 2.5% 0 15% 0;
-    }
-
-    .detail_body h3 {
-        width: fit-content;
-        margin: 2.5% 0 2.5% 0;
-        display: inline;
-        box-shadow: inset 0px -15px 0 #7BC8E04D;
-    }
-
-    .detail_body p {
-        margin-top: 1%;
-        margin-left: 5%;
-        width: 50%;
-        text-align: left;
-        font-size: 13px;
-    }
-
-    .detail_tab button {
-        border: none;
-        background-color: transparent;
-        box-shadow: none;
-        cursor: pointer;
-        width: 10%;
-        height: 105%;
-        font-weight: 500;
-        text-align: center;
-    }
-
-    .detail_tab button.active {
-        border-top: 2.5px solid #7BC8E0;
-        border-bottom: 2.5px solid #7BC8E0;
-        border-radius: 2px;
-    }
-
-    .review_body {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-}
