--- src/menu/detail/review_write/_component/reviewWrite.css
@@ -1,74 +1,0 @@
-.title {
-    width: 380px;
-    height: 40px;
-    text-align: center;
-    font-family: Noto Sans KR;
-    font-size: 24px;
-    font-weight: 600;
-    line-height: 34.75px;
-    margin: auto;
-    margin-bottom: 30px;
-    margin-top: 39px;
-}
-button {
-    width: 433px;
-    height: 50px;
-    border: 2px;
-    border-radius: 10px;
-    background: #ffffff;
-    box-shadow: 0px 0px 4px 0px #0000004d;
-    margin: 12px auto;
-    font-weight: 600;
-    font-family: Noto Sans KR;
-    display: block;
-    text-align: left;
-    padding: 0;
-}
-
-.button-active {
-    background-color: #60c1c31a;
-    border: 1.5px solid #60c1c3;
-    color: #60c1c3;
-}
-.button_text {
-    display: inline-block;
-    vertical-align: middle;
-}
-.img {
-    width: 20px;
-    height: 20spx;
-    text-align: left;
-    margin-right: 9px;
-    margin-left: 21px;
-    vertical-align: middle;
-}
-.submit-active {
-    width: 90px;
-    height: 40px;
-    border-radius: 5px;
-    background: #60c1c3;
-    font-family: Noto Sans KR;
-    font-size: 16px;
-    font-weight: 700;
-    text-align: center;
-    color: #ffffff;
-    border-color: #60c1c3;
-    margin-top: 30px;
-}
-.submit {
-    width: 90px;
-    height: 40px;
-    border-radius: 5px;
-    font-family: Noto Sans KR;
-    font-size: 16px;
-    font-weight: 700;
-    text-align: center;
-    margin-top: 30px;
-    background: #9c9c9c4d;
-    color: #52555b;
-    border-color: #9c9c9c4d;
-}
-
-.div_style {
-    display: inline-block;
-}
