--- src/menu/login/component/LoginLogo.css
@@ -1,44 +1,0 @@
-@media screen and (min-width: 769px) {
-    .logo {
-        width: 30%;
-        justify-content: center;
-        display: grid;
-        margin: auto;
-        margin-top: 70px;
-    }
-
-    .mini_logo {
-        width: 30%;
-        font-family: Inter;
-        font-size: 21px;
-        font-weight: 700;
-        line-height: 24.2px;
-        color: #7BC8E0;
-        display: grid;
-        margin: 0 auto;
-        margin-bottom: 3%;
-        justify-content: center;
-    }
-}
-
-@media screen and (max-width: 769px) {
-    .logo {
-        width: 30%;
-        justify-content: center;
-        display: grid;
-        margin: auto;
-        margin-top: 60px;
-    }
-
-    .mini_logo {
-        width: 20%;
-        font-family: Inter;
-        font-size: 13px;
-        font-weight: 700;
-        line-height: 24.2px;
-        color: #7BC8E0;
-        display: grid;
-        margin: 0 auto 0;
-        justify-content: center;
-    }
-}
