--- src/menu/login/index.module.css
@@ -1,5 +1,0 @@
-.wrapper {
-    height: auto;
-    min-height: 100%;
-    padding-bottom: 150px;
-}
