+++ src/pages/BranchSmall.jsx
@@ -1,0 +1,9 @@
+import BranchSmall from '../component/smallClub/BranchSmall';
+
+export default function BranchSmallPage() {
+    return (
+        <>
+            <BranchSmall />
+        </>
+    );
+}
