+++ src/pages/SearchPage.jsx
@@ -1,0 +1,9 @@
+import Search from '../component/search/search';
+
+export default function SearchPage() {
+    return (
+        <>
+            <Search />
+        </>
+    );
+}
